cleanup test suite code structure
