+++ tests/fixtures/system/fixtures.json
@@ -1,5 +1,5 @@
 {
-  "force": false,
+  "force": true,
   "teardown": false,
   "service_accounts": [
     {
