+++ tests/system/ae5_tools/conftest.py
@@ -1,11 +1,8 @@
-import json
-import os
-
 import pytest
 
 from ae5_tools.api import AEAdminSession, AEUserSession
-from tests.system.common import _get_account
-from tests.utils import _get_vars
+from tests.adsp.common.utils import _get_vars
+from tests.system.state import load_account
 
 
 # Expectations: the user AE5_USERNAME should have at least three projects:
@@ -18,7 +15,7 @@ from tests.utils import _get_vars
 @pytest.fixture(scope="session")
 def user_session():
     hostname: str = _get_vars("AE5_HOSTNAME")
-    local_account: dict = _get_account(id="1")
+    local_account: dict = load_account(id="1")
     username: str = local_account["username"]
     password: str = local_account["password"]
