+++ tests/system/ae5_tools/test_options.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.utils import _cmd
+from tests.adsp.common.utils import _cmd
 
 
 @pytest.fixture(scope="module")
